# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:3000
...
basePath: /
...
securityDefinitions:
  AdminAuth:
    type: basic
  UserAuth:
    type: basic  
  SupplierAuth:
    type: basic
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /user:
    x-swagger-router-controller: user
    get:
      description: get the users' list
      parameters:
        - in: query
          name: top10
          required: false
          type: boolean
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetUsersListResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        "400":
          description: bad request
          schema:
            $ref: "#/definitions/badRequestResponse"
        "500":
          description: internal server error
          schema:
            $ref: "#/definitions/internalServerErrorRequestResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  
    post:
      description: add a new user to the list
      parameters:
        - name: user
          description:  user information
          in: body
          required: true
          schema:
            $ref: "#/definitions/user"
      responses:
        "201":
          description: created
          schema:
            $ref: "#/definitions/GetuserResponse"
        "400":
          description: bad request
          schema:
            $ref: "#/definitions/badRequestResponse"
        "500":
          description: internal server error
          schema:
            $ref: "#/definitions/internalServerErrorRequestResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      security: []
  /user/{id}:
    x-swagger-router-controller: user
    get:
      description: get infromation about one user
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetuserResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        "500":
          description: internal server error
          schema:
            $ref: "#/definitions/internalServerErrorRequestResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
         - AdminAuth: []
         - UserAuth: []
    put:
      description: update an existing user
      # define the parameters
      parameters:
        - name: id
          description: user id
          type: string
          in: path
          required: true
        - name: user
          description: user
          in: body
          required: true
          schema:
            $ref: "#/definitions/user"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetuserResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        "400":
          description: bad request
          schema:
            $ref: "#/definitions/badRequestResponse"
        "500":
          description: internal server error
          schema:
            $ref: "#/definitions/internalServerErrorRequestResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
         - AdminAuth: []
         - UserAuth: []
            
    delete:
      description: delete a user
      # define the parameters
      parameters:
        - name: id
          description: user id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralDeletedddSuccessfullyResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/notFoundResponse"
        "500":
          description: internal server error
          schema:
            $ref: "#/definitions/internalServerErrorRequestResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
         - AdminAuth: []
         - UserAuth: []
...
definitions:
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  UnauthorizedResponse:
    required:
      - message
    properties:
      message:
        type: string
  badRequestResponse:
    required:
      - message
    properties:
      message:
        type: string
  internalServerErrorRequestResponse:
    required:
      - message
    properties:
      message:
        type: string
  notFoundResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  GeneralDeletedddSuccessfullyResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
  GetUsersListResponse:
    properties:
      user: 
        type: object
        properties:
             _links: 
                type: object
                properties:
                    self:
                      type: object
                      properties:
                        href:
                          type: string
             _embedded:
              type: object
              properties:
                 usersList:
                    type: array
                    items:
                      type: object
                      properties:
                          _links: 
                              type: object
                              properties:
                                  self:
                                    type: object
                                    properties:
                                      href:
                                        type: string
                          _id:
                            type: string
                          username:
                            type: string
  GetuserResponse:
    type: object
    properties:
      _links:
        type: object
        properties:
          self:
            type: object
            properties:
              href:
                type: string
      _embedded:
        type: object
        properties:
          orders:
            type: array
            items:
                type: object
                properties:
                  _links:
                    type: object
                    properties:
                      self:
                        type: object
                        properties:
                          href:
                            type: string
                  _id: 
                    type: string
          totalSales:
            type: number
      _id:
        type: string
        description: user unique id
      username:
        type: string
        description: user name
      password:
        type: string
        description: user password
      email:
        type: string
        description: user email
      createdAt:
       type: string
      updatedAt:
       type: string
      __v:
       type: number
  user:
    type: object
    properties:
      username:
        type: string
        description: user name
      password:
        type: string
        description: user password
      email:
        type: string
        description: user email
    required:
      - username
      - password
      - email  
...